fix bug in quick install bash and initialize processor's database (#276)

+++ weevent-build/package.sh
@@ -6,12 +6,12 @@
 # 3. java 1.8
 # 4. nodejs 10.16
 ################################################################################
-
+weevent_version=1.1.0
 function usage(){
     echo "Usage:"
-    echo "    package master: ./package.sh --version 1.0.0"
-    echo "    package tag: ./package.sh --tag v1.0.0 --version 1.0.0"
-    echo "    package local: ./package.sh --tag local --version 1.0.0"
+    echo "    package master: ./package.sh --version ${weevent_version}"
+    echo "    package tag: ./package.sh --tag v${weevent_version} --version ${weevent_version}"
+    echo "    package local: ./package.sh --tag local --version ${weevent_version}"
 }
 
 version=""
@@ -196,7 +196,7 @@ function tar_weevent(){
     for commonjar in $(ls ${out_path}/modules/broker/lib/);
     do
         # copy common jar into modules lib
-        if [[ -e ${out_path}/modules/governance/lib/${commonjar} ]]; then
+        if [[ (-e ${out_path}/modules/governance/lib/${commonjar}) && (-e ${out_path}/modules/processor/lib/${commonjar}) ]]; then
             cp ${out_path}/modules/broker/lib/${commonjar} ${out_path}/modules/lib
             rm ${out_path}/modules/governance/lib/${commonjar}
             rm ${out_path}/modules/processor/lib/${commonjar}
